tests/wpt/tests/svg/painting/parsing/stroke-dashoffset-computed.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: getComputedStyle().strokeDashoffset</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#StrokeDashoffsetProperty"/>
    <h:meta name="assert" content="stroke-dashoffset computed value is absolute length."/>
  </metadata>
  <g id="target"></g>
  <style>
    #target {
      font-size: 40px;
    }
  </style>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/computed-testcommon.js"/>
  <script><![CDATA[

test_computed_value("stroke-dashoffset", "10", "10px");
test_computed_value("stroke-dashoffset", "0.5em", "20px");
test_computed_value("stroke-dashoffset", "calc(10px + 0.5em)", "30px");
test_computed_value("stroke-dashoffset", "calc(10px - 0.5em)", "-10px");
test_computed_value("stroke-dashoffset", "-40%");
test_computed_value("stroke-dashoffset", "calc(50% + 60px)");

// https://www.w3.org/TR/css-values-3/#absolute-lengths
test_computed_value("stroke-dashoffset", "254cm", "9600px");
test_computed_value("stroke-dashoffset", "2540mm", "9600px");
test_computed_value("stroke-dashoffset", "10160Q", "9600px");
test_computed_value("stroke-dashoffset", "1in", "96px");
test_computed_value("stroke-dashoffset", "6pc", "96px");
test_computed_value("stroke-dashoffset", "72pt", "96px");

  ]]></script>
</svg>
